fix user

+++ dto/userDTO.go
@@ -2,7 +2,7 @@ package dto
 
 //UserUpdateDTO is used by client when PUT update profile
 type UserUpdateDTO struct {
-	UserId      int64  `json:"user_id"`
+	UserId      int    `json:"user_id"`
 	Name        string `json:"name"  binding:"required"`
 	PhoneNumber string `json:"phone_number"  binding:"required"`
 	Email       string `json:"email"  binding:"required"`
